+++ fakeredis/_msgs.py
@@ -147,6 +147,9 @@ TIMESERIES_BAD_FILTER_EXPRESSION = "TSDB: failed parsing labels"
 HEXPIRE_NUMFIELDS_DIFFERENT = "The `numfields` parameter must match the number of arguments"
 HEXPIRE_INVALID_TIME_MSG = "ERR invalid expire time, must be >= 0"
 
+# Dragonfly specific error messages
+THROTTLE_ZERO_RATES_MSG = "ERR zero rates are not supported"
+
 MISSING_ACLFILE_CONFIG = "ERR This Redis instance is not configured to use an ACL file. You may want to specify users via the ACL SETUSER command and then issue a CONFIG REWRITE (assuming you have a Redis configuration file set) in order to store users in the Redis configuration."
 
 NO_PERMISSION_ERROR = "NOPERM User {} has no permissions to run the '{}' command"